countavx512_32.s: hot fix clobbered registers in accum16
This is a mistake introduced by the improved kernel routine:
ZMM0 to ZMM3 used to be free to use when accum16 was called
due to possible overflow in the counters, but now that some
bits are carried over between iterations, this is no longer
the case.  Switch to other registers to fix the code.

Will fix this problem upstream shortly.